+++ src/NLU.DevOps.Luis/LuisClient.cs
@@ -5,14 +5,9 @@ namespace NLU.DevOps.Luis
 {
     using System;
     using System.Collections.Generic;
-    using System.Linq;
-    using System.Net;
-    using System.Net.Http;
-    using System.Text;
     using System.Threading;
     using System.Threading.Tasks;
     using Logging;
-    using Microsoft.Azure.CognitiveServices.Language.LUIS.Authoring;
     using Microsoft.Azure.CognitiveServices.Language.LUIS.Authoring.Models;
     using Microsoft.Azure.CognitiveServices.Language.LUIS.Runtime;
     using Microsoft.Azure.CognitiveServices.Language.LUIS.Runtime.Models;
@@ -24,10 +19,11 @@ namespace NLU.DevOps.Luis
 
     internal class LuisClient : ILuisClient
     {
-        private const string Protocol = "https://";
-        private const string Domain = ".api.cognitive.microsoft.com";
+        public const string Protocol = "https://";
+        public const string Domain = ".api.cognitive.microsoft.com";
 
         private static readonly TimeSpan ThrottleQueryDelay = TimeSpan.FromMilliseconds(100);
+        private readonly ILuisTrainingClient trainingClient;
 
         public LuisClient(
             string authoringKey,
@@ -37,32 +33,16 @@ namespace NLU.DevOps.Luis
             AzureSubscriptionInfo azureSubscriptionInfo,
             bool isStaging)
         {
-            this.IsStaging = isStaging;
+            if (authoringKey != null && authoringRegion != null)
+            {
+                this.trainingClient = new LuisTrainingClient(authoringKey, authoringRegion, azureSubscriptionInfo, isStaging);
+            }
 
-            var endpointOrAuthoringKey = endpointKey ?? authoringKey ?? throw new ArgumentException($"Must specify either '{nameof(authoringKey)}' or '{nameof(endpointKey)}'.");
-            this.EndpointRegion = endpointRegion ?? authoringRegion ?? throw new ArgumentException($"Must specify either '{nameof(authoringRegion)}' or '{nameof(endpointRegion)}'.");
-            this.AzureSubscriptionInfo = azureSubscriptionInfo;
-            this.AuthoringKey = authoringKey;
-
-            var endpointCredentials = new Microsoft.Azure.CognitiveServices.Language.LUIS.Runtime.ApiKeyServiceClientCredentials(endpointOrAuthoringKey);
+            var endpointCredentials = new ApiKeyServiceClientCredentials(endpointKey);
             this.RuntimeClient = new LUISRuntimeClient(endpointCredentials)
             {
-                Endpoint = $"{Protocol}{this.EndpointRegion}{Domain}",
+                Endpoint = $"{Protocol}{endpointRegion}{Domain}",
             };
-
-            this.LazyAuthoringClient = new Lazy<LUISAuthoringClient>(() =>
-            {
-                if (authoringKey == null || authoringRegion == null)
-                {
-                    throw new InvalidOperationException("Must provide authoring key and region to perform authoring operations.");
-                }
-
-                var authoringCredentials = new Microsoft.Azure.CognitiveServices.Language.LUIS.Authoring.ApiKeyServiceClientCredentials(authoringKey);
-                return new LUISAuthoringClient(authoringCredentials)
-                {
-                    Endpoint = $"{Protocol}{authoringRegion}{Domain}",
-                };
-            });
 
             this.LazySpeechConfig = new Lazy<SpeechConfig>(() => SpeechConfig.FromSubscription(endpointKey, endpointRegion));
         }
@@ -71,68 +51,38 @@ namespace NLU.DevOps.Luis
 
         private static Lazy<ILogger> LazyLogger { get; } = new Lazy<ILogger>(() => ApplicationLogger.LoggerFactory.CreateLogger<LuisNLUService>());
 
-        private string EndpointRegion { get; }
-
-        private AzureSubscriptionInfo AzureSubscriptionInfo { get; }
-
-        private string AuthoringKey { get; }
-
-        private bool IsStaging { get; }
-
         private LUISRuntimeClient RuntimeClient { get; }
-
-        private LUISAuthoringClient AuthoringClient => this.LazyAuthoringClient.Value;
-
-        private Lazy<LUISAuthoringClient> LazyAuthoringClient { get; }
 
         private Lazy<SpeechConfig> LazySpeechConfig { get; }
 
-        public async Task<string> CreateAppAsync(string appName, CancellationToken cancellationToken)
+        public Task<string> CreateAppAsync(string appName, CancellationToken cancellationToken)
         {
-            var request = new ApplicationCreateObject
-            {
-                Name = appName,
-                Culture = "en-us",
-            };
-
-            // Creating LUIS app.
-            var appId = await this.AuthoringClient.Apps.AddAsync(request, cancellationToken).ConfigureAwait(false);
-
-            // Assign Azure resource to LUIS app.
-            if (this.AzureSubscriptionInfo != null)
-            {
-                await this.AssignAzureResourceAsync(appId).ConfigureAwait(false);
-            }
-
-            return appId.ToString();
+            return this.trainingClient?.CreateAppAsync(appName, cancellationToken);
         }
 
         public Task DeleteAppAsync(string appId, CancellationToken cancellationToken)
         {
-            return this.AuthoringClient.Apps.DeleteAsync(Guid.Parse(appId), cancellationToken);
+            return this.trainingClient?.DeleteAppAsync(appId, cancellationToken);
         }
 
-        public async Task<IEnumerable<string>> GetTrainingStatusAsync(string appId, string versionId, CancellationToken cancellationToken)
+        public Task<IEnumerable<string>> GetTrainingStatusAsync(string appId, string versionId, CancellationToken cancellationToken)
         {
-            IList<ModelTrainingInfo> modelTrainingInfos = await this.AuthoringClient.Train.GetStatusAsync(Guid.Parse(appId), versionId, cancellationToken).ConfigureAwait(false);
-            return modelTrainingInfos.Select(modelInfo => modelInfo.Details.Status);
+            return this.trainingClient?.GetTrainingStatusAsync(appId, versionId, cancellationToken);
         }
 
         public Task ImportVersionAsync(string appId, string versionId, LuisApp luisApp, CancellationToken cancellationToken)
         {
-            return this.AuthoringClient.Versions.ImportAsync(Guid.Parse(appId), luisApp, versionId, cancellationToken);
+            return this.trainingClient?.ImportVersionAsync(appId, versionId, luisApp, cancellationToken);
         }
 
         public Task PublishAppAsync(string appId, string versionId, CancellationToken cancellationToken)
         {
-            var request = new ApplicationPublishObject
-            {
-                IsStaging = this.IsStaging,
-                Region = this.EndpointRegion,
-                VersionId = versionId,
-            };
+            return this.trainingClient?.PublishAppAsync(appId, versionId, cancellationToken);
+        }
 
-            return this.AuthoringClient.Apps.PublishAsync(Guid.Parse(appId), request, cancellationToken);
+        public Task TrainAsync(string appId, string versionId, CancellationToken cancellationToken)
+        {
+            return this.trainingClient?.TrainAsync(appId, versionId, cancellationToken);
         }
 
         public async Task<LuisResult> QueryAsync(string appId, string text, CancellationToken cancellationToken)
@@ -183,11 +133,6 @@ namespace NLU.DevOps.Luis
             }
         }
 
-        public Task TrainAsync(string appId, string versionId, CancellationToken cancellationToken)
-        {
-            return this.AuthoringClient.Train.TrainVersionAsync(Guid.Parse(appId), versionId, cancellationToken);
-        }
-
         public void Dispose()
         {
             this.Dispose(true);
@@ -198,32 +143,12 @@ namespace NLU.DevOps.Luis
         {
             if (disposing)
             {
-                using (this.AuthoringClient)
+                (this.trainingClient as IDisposable)?.Dispose();
+
                 using (this.RuntimeClient)
                 {
                 }
             }
-        }
-
-        private async Task AssignAzureResourceAsync(Guid appId)
-        {
-            var jsonBody = JsonConvert.SerializeObject(this.AzureSubscriptionInfo);
-            var data = new StringContent(jsonBody, Encoding.UTF8, "application/json");
-            var url = $"{this.AuthoringClient.Endpoint}/luis/api/v2.0/apps/{appId}/azureaccounts";
-            var request = new HttpRequestMessage
-            {
-                Method = HttpMethod.Post,
-                RequestUri = new Uri(url),
-                Headers =
-                {
-                    { HttpRequestHeader.Authorization.ToString(), $"Bearer {this.AzureSubscriptionInfo.ArmToken}" },
-                    { "Ocp-Apim-Subscription-Key", this.AuthoringKey }
-                },
-                Content = data,
-            };
-
-            var result = await this.AuthoringClient.HttpClient.SendAsync(request).ConfigureAwait(false);
-            result.EnsureSuccessStatusCode();
         }
     }
 }
